introduced mock data module and integrated with components
